Added a call to the runner inside game_end implementation. (Void Stranger hack)
This is needed by Void Stranger game to save without corrupting save file. Just released an optimization patch to make posible to run it on Vita along a modified standalone YoYoLoader vpk but wanted to make it posible play it with regular app. Still have to add it to the compatibility list.
Don't know if this is the best way to make it, as maybe this should be called if platTarget is set to Windows so GameEnd event is triggered before exit as it does on desktop platforms. But for the case just mimicked the deltarune_hack thing.